Mon, Feb 20, 2023, 3:33 PM +01:00

+++ credit/credit.c
@@ -72,19 +72,19 @@ bool checksum(long card_num)
     int sum2 = 0;
     for (int i = 0; i < length; i++)
     {
-        long digit = 0;
+        int digit = 0;
         if (i % 2 == 0)
         {
             digit = card_num % 10;
             card_num = card_num / 10;
-            printf("%li\n", digit);
+            printf("%i\n", digit);
             sum1 += digit * 2;
         }
         else
         {
             digit = card_num % 10;
             card_num = card_num / 10;
-            printf("%li\n", digit);
+            printf("%i\n", digit);
             sum1 += digit;
         }
     }
